api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["bjablonski <bjablonski@fg-it.eu>"]

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.13.1"
Django = "^4.0.6"
django-admin = "^2.0.1"
djoser = "^2.1.0"
black = "^22.6.0"
django-cors-headers = "^3.13.0"
pyyaml = "^6.0"
uritemplate = "^4.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
